Разделил контроллеры: Авторизации и Регистрации
