+++ src/main/java/ru/skypro/homework/controller/RegController.java
@@ -1,0 +1,40 @@
+package ru.skypro.homework.controller;
+
+
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import io.swagger.v3.oas.annotations.tags.Tag;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
+import ru.skypro.homework.dto.Register;
+import ru.skypro.homework.service.RegService;
+
+@Slf4j
+@CrossOrigin(value = "http://localhost:3000")
+@RestController
+@RequiredArgsConstructor
+@Tag(name = "Регистрация", description = "API для регистрации пользователей")
+public class RegController {
+
+    private final RegService regService;
+
+    @Operation(
+            summary = "Регистрация пользователя",
+            description = "Создает нового пользователя в системе"
+    )
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "201", description = "Пользователь успешно создан"),
+            @ApiResponse(responseCode = "400", description = "Некорректные данные или пользователь уже существует")
+    })
+    @PostMapping("/register")
+    @ResponseStatus(HttpStatus.CREATED)
+    public void register(@RequestBody Register register) {
+        if (!regService.register(register)) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+        }
+    }
+}
