+++ src/main/java/ru/skypro/homework/service/RegService.java
@@ -1,0 +1,9 @@
+package ru.skypro.homework.service;
+
+import ru.skypro.homework.dto.Register;
+
+public interface RegService {
+
+    boolean register(Register register);
+
+}
