+++ src/main/java/ru/skypro/homework/service/impl/RegServiceImpl.java
@@ -1,0 +1,36 @@
+package ru.skypro.homework.service.impl;
+
+import org.springframework.security.core.userdetails.User;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.security.provisioning.UserDetailsManager;
+import org.springframework.stereotype.Service;
+import ru.skypro.homework.dto.Register;
+import ru.skypro.homework.service.RegService;
+
+@Service
+public class RegServiceImpl implements RegService {
+
+    private final UserDetailsManager manager;
+    private final PasswordEncoder encoder;
+
+    public RegServiceImpl(UserDetailsManager manager, PasswordEncoder encoder) {
+        this.manager = manager;
+        this.encoder = encoder;
+    }
+
+    @Override
+    public boolean register(Register register) {
+        if (manager.userExists(register.getUsername())) {
+            return false;
+        }
+        manager.createUser(
+                User.builder()
+                        .passwordEncoder(this.encoder::encode)
+                        .password(register.getPassword())
+                        .username(register.getUsername())
+                        .roles(register.getRole().name())
+                        .build());
+        return true;
+    }
+
+}
